requirements.txt
alembic==1.15.2
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
coverage==7.6.10
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.7
decorator==5.1.1
docker-pycreds==0.4.0
elasticai.creator @ git+ssh://git@github.com/es-ude/elastic-ai.creator.git@e7e8769b4012e2907b9e428df1854b38abe950b3
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multimethod==1.12
mypy==1.14.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optuna==4.2.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visions==0.7.6
wandb==0.18.6
wcwidth==0.2.13
wordcloud==1.9.3
ydata-profiling==4.12.0
